RSPBA scraper that works for 2022
This uses selenium to go directly to the 2022 results page and gets all the contests.
For each contest, get the drop down for contest, then get the appropriate table (find the right div then get the table)
Put the table in a list and at the end, dump the list to a CSV file.
I still want to improve the data storage format. Would be nice to save/load from a DB, JSON or something else.
I also need to add the dates for the contest.